multion_agent/requirements.txt
firebase-functions
firebase-admin
crewai
crewai_tools
langchain_groq
python-dotenv
